requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
email-validator==1.1.3
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
psycopg2-binary==2.9.10
gunicorn==20.1.0  # For production deployment
stripe==2.60.0  # For payment processing
bcrypt==4.0.1  # For password hashing
PyJWT==2.3.0  # For JWT tokens
itsdangerous==2.2.0  # For secure signing
Jinja2==3.1.5  # Template engine
MarkupSafe==3.0.2
alembic==1.14.1  # For database migrations
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
greenlet==3.1.1
click==8.1.8
blinker==1.9.0
colorama==0.4.6
Mako==1.3.9
typing_extensions==4.12.2 